[refactor] Burst the primer command in three files
